[FIX] 주문 테이블에 총 주문 가격 추가

+++ src/test/java/com/dailyon/orderservice/domain/torder/service/TOrderServiceTest.java
@@ -246,7 +246,7 @@ class TOrderServiceTest extends ContainerBaseTestSupport {
     // when then
     String orderId = tOrderService.createTOrder(request, memberId).getId();
     TOrder tOrder = orderDynamoRepository.findById(orderId).get();
-    Integer totalAmount = tOrder.calculateTotalAmount();
+    Long totalAmount = tOrder.calculateTotalAmount();
     assertThat(totalAmount).isEqualTo(49900000);
   }
